# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: G.G. Maidan Esports Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tournaments:
    get:
      operationId: listTournaments
      tags: [tournaments]
      summary: List tournaments
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [upcoming, live, completed]
        - name: gameId
          in: query
          schema:
            type: ["integer", "null"]
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: List of tournaments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Tournament"
    post:
      operationId: createTournament
      tags: [tournaments]
      summary: Create a tournament
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TournamentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tournament"
  /tournaments/{id}:
    get:
      operationId: getTournament
      tags: [tournaments]
      summary: Get tournament details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Tournament details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TournamentDetail"
        "404":
          description: Not found
  /tournaments/{id}/standings:
    get:
      operationId: getTournamentStandings
      tags: [tournaments]
      summary: Get tournament standings
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Standings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Standing"
  /tournaments/{id}/register:
    post:
      operationId: registerForTournament
      tags: [tournaments]
      summary: Register a team for a tournament
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegistrationInput"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Registration"
...
components:
  schemas:
...
    Tournament:
      type: object
      required: [id, title, gameName, status, startDate, prizePool, maxTeams, registeredTeams, bannerUrl]
      properties:
        id:
          type: integer
        title:
          type: string
        gameId:
          type: integer
        gameName:
          type: string
        status:
          type: string
          enum: [upcoming, live, completed]
        startDate:
          type: string
        endDate:
          type: ["string", "null"]
        prizePool:
          type: string
        maxTeams:
          type: integer
        registeredTeams:
          type: integer
        bannerUrl:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
    TournamentDetail:
      type: object
      required: [id, title, gameName, status, startDate, prizePool, maxTeams, registeredTeams, bannerUrl, rules, schedule, brackets]
      properties:
        id:
          type: integer
        title:
          type: string
        gameId:
          type: integer
        gameName:
          type: string
        status:
          type: string
          enum: [upcoming, live, completed]
        startDate:
          type: string
        endDate:
          type: ["string", "null"]
        prizePool:
          type: string
        maxTeams:
          type: integer
        registeredTeams:
          type: integer
        bannerUrl:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        rules:
          type: ["string", "null"]
        schedule:
          type: array
          items:
            $ref: "#/components/schemas/ScheduleItem"
        brackets:
          type: array
          items:
            $ref: "#/components/schemas/BracketMatch"
    TournamentInput:
      type: object
      required: [title, gameId, startDate, prizePool, maxTeams]
      properties:
        title:
          type: string
        gameId:
          type: integer
        startDate:
          type: string
        endDate:
          type: ["string", "null"]
        prizePool:
          type: string
        maxTeams:
          type: integer
        bannerUrl:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        rules:
          type: ["string", "null"]
    ScheduleItem:
      type: object
      required: [id, label, datetime, description]
      properties:
        id:
          type: integer
        label:
          type: string
        datetime:
          type: string
        description:
          type: ["string", "null"]
    BracketMatch:
      type: object
      required: [id, round, team1Name, team2Name, winner, score1, score2]
      properties:
        id:
          type: integer
        round:
          type: integer
        team1Name:
          type: ["string", "null"]
        team2Name:
          type: ["string", "null"]
        winner:
          type: ["string", "null"]
        score1:
          type: ["integer", "null"]
        score2:
          type: ["integer", "null"]
    Standing:
      type: object
      required: [rank, teamId, teamName, teamLogoUrl, kills, points, placement]
      properties:
        rank:
          type: integer
        teamId:
          type: integer
        teamName:
          type: string
        teamLogoUrl:
          type: ["string", "null"]
        kills:
          type: integer
        points:
          type: integer
        placement:
          type: integer
    RegistrationInput:
      type: object
      required: [teamId]
      properties:
        teamId:
          type: integer
        playerIds:
          type: array
          items:
            type: integer
    Registration:
      type: object
      required: [id, tournamentId, teamId, registeredAt, status]
      properties:
        id:
          type: integer
        tournamentId:
          type: integer
        teamId:
          type: integer
        registeredAt:
          type: string
        status:
          type: string
...