# lib/api-zod/src/generated/types/registrationInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * G.G. Maidan Esports Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface RegistrationInput {
  teamId: number;
  playerIds?: number[];
}